Bump `0.2.0` to reflect semver-incompatible change
As mentioned in #17, #11 among other things changed the public API in
a way that causes failures to build.

Thus those changes should be released under the `0.2.0` version to not
break existing or new projects that don't lock the version to `0.1.1`.

To close #17, and stop churn for the ecosystem `0.1.2` would need to be
yanked as well.